+++ shex_validation/src/lib.rs
@@ -9,6 +9,8 @@ pub mod reason;
 pub mod rule;
 pub mod schema_without_imports;
 pub mod schema_without_imports_error;
+pub mod shex_config;
+pub mod shex_format;
 pub mod solver;
 pub mod validator;
 pub mod validator_config;
@@ -22,6 +24,8 @@ pub use crate::result_value::*;
 pub use crate::rule::*;
 pub use crate::schema_without_imports::*;
 pub use crate::schema_without_imports_error::*;
+pub use crate::shex_config::*;
+pub use crate::shex_format::*;
 pub use crate::validator::*;
 pub use crate::validator_config::*;
 pub use crate::validator_error::*;
@@ -30,13 +34,6 @@ pub use crate::validator_runner::*;
 /// Default MAX STEPS
 /// This value can be overriden in the Validator configuration
 const MAX_STEPS: usize = 20;
-
-/// Contains possible ShEx formats
-#[derive(Debug, PartialEq, Clone)]
-pub enum ShExFormat {
-    ShExC,
-    ShExJ,
-}
 
 /// Method employed to resolve imports when ghessing the format of an import
 #[derive(Debug, Clone)]