+++ shex_validation/src/shex_format.rs
@@ -1,0 +1,8 @@
+use serde_derive::{Deserialize, Serialize};
+
+/// Contains possible ShEx formats
+#[derive(Deserialize, Serialize, Debug, PartialEq, Clone)]
+pub enum ShExFormat {
+    ShExC,
+    ShExJ,
+}